finalproject/requirements.txt
streamlit
opencv-python
numpy
Pillow
deepface
google-generativeai
